requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
dj-rest-auth==3.0.0
    # via django-react-oauth (setup.py)
django==4.1.7
    # via
    #   dj-rest-auth
    #   django-react-oauth (setup.py)
    #   djangorestframework
    #   djangorestframework-simplejwt
django-webpack-loader==1.8.1
    # via django-react-oauth (setup.py)
djangorestframework==3.14.0
    # via
    #   dj-rest-auth
    #   django-react-oauth (setup.py)
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via django-react-oauth (setup.py)
gunicorn==20.1.0
    # via django-react-oauth (setup.py)
psycopg2==2.9.5
    # via django-react-oauth (setup.py)
pydantic==1.10.5
    # via django-react-oauth (setup.py)
pyjwt==2.6.0
    # via djangorestframework-simplejwt
python-dotenv==1.0.0
    # via django-react-oauth (setup.py)
pytz==2022.7.1
    # via djangorestframework
sqlparse==0.4.3
    # via django
typing-extensions==4.5.0
    # via pydantic

# The following packages are considered to be unsafe in a requirements file:
# setuptools
